Fixes import
Fix the coords update for import display
Lot of todos on this one about the element selection

+++ assets/scripts/src/bb.gmap.controller.js
@@ -433,7 +433,10 @@ BB.gmap.controller.prototype.listeners = function()
         	case 46:
         		if (that.focused()) {
         			// Remove focused item
-        			that.focused().delete();
+        			if (that.focused().data('editable')) {
+	        			that.focused().delete();
+        				that.remove_focus();
+	        		}
         		}
         	break;
